+++ gimel-dataapi/gimel-sql/src/main/scala/com/paypal/gimel/sql/GimelQueryProcessor.scala
@@ -31,10 +31,11 @@ import org.apache.spark.streaming.{Seconds, StreamingContext}
 import com.paypal.gimel._
 import com.paypal.gimel.common.catalog.{CatalogProvider, DataSetProperties}
 import com.paypal.gimel.common.conf.{CatalogProviderConfigs, GimelConstants}
+import com.paypal.gimel.common.utilities.{DataSetType, DataSetUtils}
 import com.paypal.gimel.common.utilities.Timer
 import com.paypal.gimel.datasetfactory.GimelDataSet
 import com.paypal.gimel.datastreamfactory.{StreamingResult, StructuredStreamingResult, WrappedData}
-import com.paypal.gimel.jdbc.conf.{JdbcConfigs, JdbcConstants}
+import com.paypal.gimel.jdbc.conf.JdbcConfigs
 import com.paypal.gimel.kafka.conf.{KafkaConfigs, KafkaConstants}
 import com.paypal.gimel.logger.Logger
 import com.paypal.gimel.logging.GimelStreamingListener
@@ -170,6 +171,12 @@ object GimelQueryProcessor {
         }
         stringToDF(sparkSession, resultingStr)
       } else {
+
+        // Set HBase Page Size for optimization if selecting from HBase with limit
+        if (QueryParserUtils.isHavingLimit(sql)) {
+          setLimitForHBase(sql, options, sparkSession)
+        }
+
         val (originalSQL, destination, selectSQL, kafkaDataSets, queryPushDownFlag) =
           resolveSQL(sql, sparkSession, dataSet)
         destination match {
